chore: rename common to utils
